src/main/java/core/basesyntax/service/handler/BalanceOperationHandler.java
package core.basesyntax.service.handler;

import core.basesyntax.db.Storage;
import core.basesyntax.model.FruitTransaction;

public class BalanceOperationHandler implements OperationHandler {
    @Override
    public void handleTransaction(FruitTransaction transaction) {
        if (transaction.getQuantity() < 0) {
            throw new IllegalArgumentException("Quantity cannot be negative");
        }
        Storage.fruits.put(transaction.getFruit(), transaction.getQuantity());
    }
}
